requirements.txt
pandas>=2.3.3
numpy>=2.4.0
matplotlib>=3.10.8
statsmodels>=0.14.6
openpyxl>=3.1.5
scipy>=1.16.3
flask>=3.1.2
python-docx>=1.2.0
blinker>=1.9.0
click>=8.1.3
itsdangerous>=2.2.0
jinja2>=3.1.2
markupsafe>=2.1.1
werkzeug>=3.1.0
lxml>=6.0.2
typing_extensions>=4.15.0
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.7
contourpy>=1.0.1
cycler>=0.10
fonttools>=4.22.0
kiwisolver>=1.3.1
packaging>=20.0
pillow>=8
pyparsing>=3
patsy>=0.5.6
et-xmlfile>=2.0.0
colorama>=0.4.6
six>=1.5